refactor: request, response 패키지 구조 변경
