+++ dailyge-api/src/main/java/project/dailyge/app/core/weeklygoal/presentation/response/WeeklyGoalCreateResponse.java
@@ -1,4 +1,4 @@
-package project.dailyge.app.core.weeklygoal.response;
+package project.dailyge.app.core.weeklygoal.presentation.response;
 
 public record WeeklyGoalCreateResponse(Long weeklyGoalId) {
