+++ dailyge-api/src/test/java/project/dailyge/app/test/weeklygoal/documentationtest/WeeklyGoalReadDocumentationTest.java
@@ -8,7 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.restdocs.restassured.RestDocumentationFilter;
 import project.dailyge.app.common.DatabaseTestBase;
 import project.dailyge.app.core.weeklygoal.application.WeeklyGoalWriteService;
-import project.dailyge.app.core.weeklygoal.request.WeeklyGoalCreateRequest;
+import project.dailyge.app.core.weeklygoal.presentation.request.WeeklyGoalCreateRequest;
 
 import java.time.DayOfWeek;
 import java.time.LocalDate;
